Refactor product directory
